+++ TheLibraryIsOpen/Controllers/MovieController.cs
@@ -23,12 +23,14 @@ namespace TheLibraryIsOpen.Controllers
     public class MovieController : Controller
     {
         private readonly MovieCatalog _mc;
+        private readonly ClientStore _cs;
 
-        public MovieController(MovieCatalog mc)
+        public MovieController(MovieCatalog mc, ClientStore cs)
         {
             _mc = mc;
+            _cs = cs;
         }
-
+        
         public async Task<IActionResult> Index()
         {
             List<Movie> lm = await _mc.GetAllMoviesDataAsync();
@@ -37,8 +39,11 @@ namespace TheLibraryIsOpen.Controllers
         }
 
         [HttpGet]
-        public IActionResult Create()
+        public async Task<IActionResult> Create()
         {
+            bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+            if (!isAdmin)
+                return Unauthorized();
             ViewData["Message"] = "Your create movie page.";
             return View();
         }
@@ -46,6 +51,9 @@ namespace TheLibraryIsOpen.Controllers
         [HttpPost]
         public async Task<IActionResult> Create(Movie movie)
         {
+            bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+            if (!isAdmin)
+                return Unauthorized();
             ViewData["Message"] = "Your create movie page.";
             await _mc.CreateMovieAsync(movie);
             await _mc.CommitAsync();
@@ -55,7 +63,9 @@ namespace TheLibraryIsOpen.Controllers
         [HttpGet]
         public async Task<ActionResult> Delete(int id)
         {
-
+            bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+            if (!isAdmin)
+                return Unauthorized();
             Movie toDelete = await _mc.GetMovieByIdAsync(id);
 
             return View(toDelete);
@@ -67,6 +77,9 @@ namespace TheLibraryIsOpen.Controllers
         {
             try
             {
+                bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+                if (!isAdmin)
+                    return Unauthorized();
                 ViewData["Message"] = "Your delete movie page.";
                 await _mc.DeleteMovieAsync(movie);
                 await _mc.CommitAsync();
@@ -93,6 +106,9 @@ namespace TheLibraryIsOpen.Controllers
         [HttpGet]
         public async Task<ActionResult> Edit(int id)
         {
+            bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+            if (!isAdmin)
+                return Unauthorized();
             Movie toEdit = await _mc.GetMovieByIdAsync(id);
             toEdit.Producers = await _mc.GetAllMovieProducerDataAsync(id);
             toEdit.Actors = await _mc.GetAllMovieActorDataAsync(id);
@@ -105,6 +121,9 @@ namespace TheLibraryIsOpen.Controllers
         {
             try
             {
+                bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+                if (!isAdmin)
+                    return Unauthorized();
                 await _mc.UpdateMovieAsync(edit.ToMovie());
                 await _mc.CommitAsync();
