+++ TheLibraryIsOpen/Controllers/MusicController.cs
@@ -1,9 +1,10 @@
-﻿using System;
+﻿using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using TheLibraryIsOpen.Controllers.StorageManagement;
 using TheLibraryIsOpen.Models.DBModels;
 
@@ -12,10 +13,12 @@ namespace TheLibraryIsOpen.Controllers
     public class MusicController : Controller
     {
         private readonly MusicCatalog _mc;
+        private readonly ClientStore _cs;
 
-        public MusicController(MusicCatalog mc)
+        public MusicController(MusicCatalog mc, ClientStore cs)
         {
             _mc = mc;
+            _cs = cs;
         }
 
         public async Task<IActionResult> Index()
@@ -40,8 +43,11 @@ namespace TheLibraryIsOpen.Controllers
             return View(music);
         }
 
-        public IActionResult Create()
+        public async Task<IActionResult> Create()
         {
+            bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+            if (!isAdmin)
+                return Unauthorized();
             return View();
         }
 
@@ -49,6 +55,9 @@ namespace TheLibraryIsOpen.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Music music)
         {
+            bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+            if (!isAdmin)
+                return Unauthorized();
             if (ModelState.IsValid)
             {
                 await _mc.CreateMusicAsync(music);
@@ -66,6 +75,10 @@ namespace TheLibraryIsOpen.Controllers
                 return NotFound();
             }
 
+            bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+            if (!isAdmin)
+                return Unauthorized();
+
             var music = await _mc.FindMusicByIdAsync(id);
             if (music == null)
             {
@@ -82,6 +95,9 @@ namespace TheLibraryIsOpen.Controllers
             {
                 return NotFound();
             }
+            bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+            if (!isAdmin)
+                return Unauthorized();
 
             if (ModelState.IsValid)
             {
@@ -114,6 +130,10 @@ namespace TheLibraryIsOpen.Controllers
                 return NotFound();
             }
 
+            bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+            if (!isAdmin)
+                return Unauthorized();
+
             var music = await _mc.FindMusicByIdAsync(id);
 
             if (music == null)
@@ -130,6 +150,11 @@ namespace TheLibraryIsOpen.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
+
+            bool isAdmin = await _cs.IsItAdminAsync(User.Identity.Name);
+            if (!isAdmin)
+                return Unauthorized();
+
             var music = await _mc.FindMusicByIdAsync(id);
             await _mc.DeleteMusicAsync(music);
             await _mc.CommitAsync();
